refactor(armory): extract armory logic completly in own addon
