+++ addons/armory/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\coopr\addons\armory
