+++ addons/armory/CfgFunctions.hpp
@@ -1,0 +1,12 @@
+class CfgFunctions {
+    class coopr {
+        class coopr_armory_functions {
+            file = "x\coopr\addons\armory\functions";
+            class showArmoryMenu {};
+            class equipHandler {};
+            class selectHandler {};
+            class setupArmoryModule {};
+            class initArmoryNPC {};
+        }
+    }
+}
