+++ addons/armory/config.cpp
@@ -1,7 +1,7 @@
 #include "script_component.hpp"
 
 class CfgPatches {
-    class coopr_equipment
+    class coopr_armory
     {
         // Meta information for editor
         name = "CoopR Mod";
@@ -13,13 +13,16 @@ class CfgPatches {
             "coopr_core"
         };
         // List of objects (CfgVehicles classes) contained in the addon. Important also for Zeus content unlocking.
-        units[] = {};
+        units[] = {
+            "CoopR_ModuleSetupArmory"
+        };
         // List of weapons (CfgWeapons classes) contained in the addon.
         weapons[] = {};
     };
 }
 
 #include "CfgFunctions.hpp"
+#include "CfgVehicles.hpp"
 // UI
 #include "\x\coopr\addons\core\define.hpp"
-#include "guiArmory.hpp"
+#include "ui\guiArmory.hpp"
