+++ addons/armory/functions/fn_setupArmoryModule.sqf
@@ -2,6 +2,9 @@
 
 params [["_logic", objNull]];
 
+private _loggingLevel = _logic getVariable ["Logging", -1];
+[_loggingLevel, DEBUG_CTX] call coopr_fnc_setLogLevel;
+
 private _syncedObjects = synchronizedObjects _logic;
 DEBUG2("Armory Module - synced units: %1", _syncedObjects);
