+++ addons/armory/functions/script_component.hpp
