+++ addons/armory/stringtable.xml
@@ -1,27 +1,27 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <Project name="CoopR Mod">
-    <Package name="equipment">
-        <Key ID="str.coopr.equipment.action.armory">
+    <Package name="armory">
+        <Key ID="str.coopr.armory.action">
             <Original>Request new equipment</Original>
             <German>Neue Ausruestung beantragen</German>
         </Key>
-        <Key ID="str.coopr.equipment.vendor.rep">
+        <Key ID="str.coopr.armory.vendor.rep">
             <Original>Your current reputation is: %1</Original>
             <German>Dein momentaner Ruf betraegt: %1</German>
         </Key>
-        <Key ID="str.coopr.equipment.vendor.remain">
+        <Key ID="str.coopr.armory.vendor.remain">
             <Original>You still need %1 reputation points</Original>
             <German>Du benoetigst noch %1 mehr Ruf</German>
         </Key>
-        <Key ID="str.coopr.equipment.vendor.repneeded">
+        <Key ID="str.coopr.armory.vendor.repneeded">
             <Original>You need %1 reputation for this item</Original>
             <German>Du benoetigst %1 fuer diesen Ausruestungsgegenstand</German>
         </Key>
-        <Key ID="str.coopr.equipment.vendor.allowed">
+        <Key ID="str.coopr.armory.vendor.allowed">
             <Original>You are allowed to receive this item</Original>
             <German>Du kannst dieses Ausruestungsgegenstand erhalten</German>
         </Key>
-        <Key ID="str.coopr.equipment.vendor.norep">
+        <Key ID="str.coopr.armory.vendor.norep">
             <Original>You do not enough reputation to receive this item</Original>
             <German>Du hast nicht genug Ruf um diese Ausruestung zu erhalten</German>
         </Key>
