+++ addons/core/constants.hpp
@@ -1,11 +1,2 @@
-#define COOPR_NPC_VENDOR "coopr_npc_vendor"
-#define COOPR_NPC_INTEL "coopr_npc_intel"
-#define COOPR_NPC_INTEL_ASSIST "coopr_npc_intel_assist"
-#define COOPR_NPC_VEHICLE "coopr_npc_vehicle"
-#define COOPR_NPC_COMMANDER "coopr_npc_commander"
 
-#define COOPR_BOX_EQUIPMENT "coopr_box_equipment"
-#define COOPR_BOX_MAGAZINES "coopr_box_magazines"
 
-//AREAS
-//#define COOPR_ESC_ZONE "coopr_esc_zone"
