+++ addons/core/helper/fn_getLoadoutForRole.sqf
