--- addons/equipment/$PBOPREFIX$
@@ -1,1 +1,0 @@
-x\coopr\addons\equipment
