--- addons/equipment/CfgFunctions.hpp
@@ -1,11 +1,0 @@
-class CfgFunctions {
-    class coopr {
-        class coopr_equipment_functions {
-            file = "x\coopr\addons\equipment\functions";
-            class showArmoryMenu {};
-            class getLoadoutForRole {};
-            class equipHandler {};
-            class selectHandler {};
-        }
-    }
-}
