+++ build.cmd
@@ -12,7 +12,7 @@ mkdir %dest%
 
 %armake64% build -p addons\persistence %dest%\character.pbo
 %armake64% build -p addons\core %dest%\core.pbo
-%armake64% build -p addons\equipment %dest%\equipment.pbo
+%armake64% build -p addons\armory %dest%\armory.pbo
 %armake64% build -p addons\integration %dest%\integration.pbo
 %armake64% build -p addons\intelligence %dest%\intelligence.pbo
 %armake64% build -p addons\lobby %dest%\lobby.pbo
