+++ build.sh
@@ -11,7 +11,7 @@ mkdir -p $DEST
 
 $ARMAKE build -p -f addons/persistence $DEST/character.pbo
 $ARMAKE build -p -f addons/core $DEST/core.pbo
-$ARMAKE build -p -f addons/equipment $DEST/equipment.pbo
+$ARMAKE build -p -f addons/armory $DEST/armory.pbo
 $ARMAKE build -p -f addons/integration $DEST/integration.pbo
 $ARMAKE build -p -f addons/intelligence $DEST/intelligence.pbo
 $ARMAKE build -p -f addons/lobby $DEST/lobby.pbo
